Create base class for template collection
